requirements.txt
clr==1.0.3
GPUtil==1.4.0
paho-mqtt==1.6.1
psutil==5.9.5
PyYAML==6.0.1
